refactor: move general helpers to utils.ts
Moved non-expect helper functions from expectUtils.ts to utils.ts:
- formatPosition
- getInventorySummary
- getEntityName
- isBlockEmpty

expectUtils.ts now only contains expect* functions that throw errors
for validation purposes.
